requirements.txt
streamlit
tensorflow==2.15
numpy
pillow
opencv-python-headless
pyshorteners
openai  # 최신 버전 사용!
dotenv